Inst Scoping unset search-realm, Refs #10733
This commit modfies specific controllers such that if these pages are browsed
when Institutional Scoping setting is on, and the user is currently scoped to
a repository (search-realm is set), they are assumed to be reverting to a
'global' context and will no longer be scoped to a specific institution.
In each of these circumstances, unset search-realm.