Inst Scoping search-realm exceptions,Refs #10733
These controllers have been changed to modify the state of the search-realm
user attribute.

The search-realm attribute is used to determine if a user is currently scoped
to a repository when the Insitutional Scoping setting is turned on. If
search-realm is set, the user will be scoped to this repo; if search-realm
is not set, the user will be globally scoped.

- actor browseAction.class.php: modified to set search-realm if the request
repos is set. If Institutional Scoping is ON, remove the user search-realm
attribute when request repos is not set.

- informationObject browseAction.class.php: when the request repos is NOT set
and Institutional Scoping is ON and the request 'collection' is NOT set,
remove the user search-realm attribute.

- search autocompleteAction.class.php: when the request repos is NOT set
and Institutional Scoping is ON, remove the user search-realm attribute.

- repository browseAction.class.php: When Institutional Scoping is turned
ON and if the request repos is set then set the user search-realm and add
the repo to the search query.  If repos is not set, remove the search-realm
attribute and do not add the repo to the query.