Inst Scoping search and repo tmplt, Refs #10733
Misc remaining changes:

- Repository _contextMenu template: modify this template to display the
Institutional Block when Institutional Scoping is ON and a repository is set.

- Global search box: When Institutional Scoping is ON, search with the top
search box is ALWAYS global. Use the Institutional Block search box to search
within an institution.

- Browse all description link: When Institutional Scoping is ON and the user
is scoped to an institution, ensure the scoped repo is part of the URL so
results are restricted to the scoped repo.

- Modify the global search box text to use the 'app_ui_label_globalSearch'
configured text.