Inst Scoping descr list realm fix, Refs# 10733
This commit corrects an issue found when testing the new Institutional
Scoping feature's Archival Description List in the sidebar. The
search-realm was not set when accessing the archival descriptions directly
from the list. This has been resolved by assigning the search-realm earlier
inside the sfIsiahPlugin.

Also added one additional change where the search-realm needed to be cleared
in the functions browse controller.