Correct event date behaviour on save, Refs #10278
An issue was reported where if modifications to existing dates were made,
and a new date was added, and then the description was saved that the new
date would be saved but the modified dates would not. This issue affects
all description templates.

On informationObject save, without any modifications to the dates,
eventsRelatedByobjectId[] would contain all events. In addition, with
modifications to one of many dates, eventsRelatedByobjectId[] would contain
all events associated with the informationObject. When there was a combination
of new records and updated ones, only the new records were contained in
eventsRelatedByobjectId[], causing only those updates to be saved.

Ensuring that eventsRelatedByobjectId[] was always populated with the date
events corrects this issue.

When using the MODS and RAD templates where an actor is also saved on the
same dialog, the same issue was occurring when there was a combination of
new actors/dates and updated actors or dates. This has also been addressed
in this fix.