Improve entity type dropdown usability for export
Previously, when objectType was set in the URL the drop down for entity type at
the export screen would auto-select the entity type the user was browsing in
the clipboard before hitting the 'export' button. However, the user could then
manually select a different entity type from the dropdown even if that entity
type had 0 clipboard items.

This commit makes it so if entityType is set, only populate the dropdown with
said entityType, not the other ones.