Add term aggregations
Mostly the same as facets with some simplifications in the populate method
and the template partial. It removes the partially introduced hability of
selecting multiple values. Some changes required in the unique records query
for the language facet as the language query is no longer added as a filter.
Drafts are now filter in the query so there is no need to do it in the aggs.