Fix i18n repository region filter, refs #11121
This fix will make it so when browsing repositories and setting the region
filter, only show regions in the dropdown for the current culture. This also
fixes repo browse so the region filter works for non-English cultures.

NOTE: This filter does not support culture fallback! You can only set the
region filter on terms for the current culture.