Remove scope for siteBaseUrl setting, refs #10309
Digital object thumbnails were not showing in item description reports because
getting the siteBaseUrl setting was failing, making the thumbnail urls invalid.

The siteBaseUrl setting would sometimes have a scope filled in (only if it was
created via the admin->settings form), which would break sfConfig::get when
trying to access the setting. This commit ensures this setting will no longer
have a scope.