Check digitalObjectPath during csv import #11448
Previously it was blindly trying to copy from a file located at
digitalObjectPath even if it didn't exist or we didn't have read perms. Now we
will display a warning to the user and gracefully skip trying to create the
digital object.

Also moved the digitalObjectPath code into its own handler function.