Add inherited creators to IOs in ES, refs #11788
Add a foreign field in the IOs in ES to store creators inherited from its
ancestors. Use a different field than creators to allow including inherited
creator names when searching over all fields but not inherited creator
histories. Like in the GUI creators are only inherited when the description
doesn't have directly related creators.

Because inherited and directly related creators are in different fields,
the IOs where the creators are inherited won't appear in the "Creator of"
list at the inherited creator index page.

Like with the 'repository' and 'partOf' fields, with the 'inheritedCreators'
field, there is a known issue when a top level is updated, the descendants
are not being updated in ES. Considering that all descendants should be
updated each time an IO changes in ES, this should be done in an
asynchronous job launched from the method that updates the IOs in ES.