Create actor unlinker CLI task, Refs# 11753
CLI task that will update cases where a creator could be linked to
information object using inheritance.

If a descr slug is specified, all creators on this node + descendants
will be examined.

If an actor slug is specified, all descrs with this creator will be
examined.
- only this actor can be modified.
- in cases where more than one creator would have been unlinked, nothing
will be done.

If actor or descr is NOT specified, all nodes will be examined to see
if creators could be inherited instead.

- corrected short description in promoteUserToAdminTask.class.php