Permission-related fix to menu items, refs #11075
This fixes a permissions-related issue preventing the "Subjects" and
"Places" links from being displayed on the home page and in the "Browse"
menu.

The permission checking logic in the menu display relies on permissions from
security.yml files. The menu items meant to direct users to the index pages
for the place and subject taxonomies had paths that linked to the taxonomy
browse action instead which would then redirect to the taxonomy index action.
Because of restrictions on the taxonomy browse action this was causing the
menu items for places and subjects not to appear in menus for
non-authenticated users.

This is fixed by:

* Fixing an issue with the taxonomy module’s index action that prevented it
  from working correctly when called with an ID as a GET parameter
* Moving access control for the taxonomy module’s index action from
  security.yml to the action logic so place and subject can be accessible to all
* Changing the paths in the menu items for the place and subject taxonomies,
  using a data migration, to so the menu paths aren’t reliant on the taxonomy
  module browse action to redirect to the taxonomy module index action

This commit also removes some unnecessary logic from the taxonomy browse action.