Permissive slug generation uri display, refs #11761
Ensure the characters * : @ = , are represented as literals in the
browser address bar when the 'Permissive Slugs' setting is activated.

- this will prevent these chars from being percent-encoded when
  'permissive slugs' setting is turned on in AtoM
- these chars are allowed in slugs when 'permissive slugs' setting is on.
- if 'permissive slugs' is OFF, use PHP's urlencode as normal.
- this should only affect the PATH segment of an AtoM URI.
- this will affect URI construction for page links and address bar.

I do not believe QubitMetadataRoute needs to be changed to extend
QubitRoute as it appears to be related to request routing which already
correctly resolves requests with either literals or percent-encoded
unicode characters in them.

While testing I discovered an issue with generating finding aids when a
slug has a semicolon in it - in the exec call to run export:bulk, the
slug needs to be enclosed in quotes to ensure characters in the slug
are not interpreted on the command line.