Remove unneeded Docker volumes, refs #12111
The 'atom_uploads' and 'atom_downloads' volumes are placed inside another
volume used to share the entire AtoM directory. This causes data loss when
the container is recreated. The entire AtoM volume is shared by the 'atom',
'nginx' and 'atom_worker' services with the same permissions than this two
volumes, so they are actually not needed.