Permissive Slug Generation fix, refs #11761
Removed modification to sfRoute and added code to override method in
QubitRoute. QubitResourceRoute is already extending QubitRoute.

I do not believe QubitMetadataRoute needs to be changed as it appears
to be related to request routing which will correctly resolve requests
with either literals or percent-encoded unicode characters in them.

In testing I also found an issue with generating finding aids when a
slug has a semicolon in it - in the exec call to run export:bulk, the
slug needs to be enclosed in quotes to ensure characters in the slug
are not interpreted on the command line.