Fix access to unescaped content
Since the `sf_escaping_strategy` may now be enabled or not, we can find
escaped or unescaped data in the templates. Therefore, we need to unescape
the data with methods that work over not escaped data. Instead of using
`->get()`, `->getRaw()` or `->getRawValue()`, that only work when they are
ran over `sfOutputEscaperArrayDecorator` and `sfOutputEscaperObjectDecorator`,
we should use `$sf_data->getRaw()` or `sfOutputEscaper::unescape()`.