Permissive slug uri display, refs #11761
The previous commit to ensure the characters * : @ = , are represented
as literals in the browser address bar when 'Permissive Slugs' setting
is activated had an issue where the class QubitMetadataRoute needs
to extend QubitRoute instead of sfRoute. This commit corrects this.

The generate() method from sfRoute also needs to be overridden in
QubitRoute to successfully call the overridden generateWithTokens()
method.