Trim CSV import elements, refs #12067
This change trims all pipe separated elements contained in an import
CSV file. This was not occurring previously and extra leading and
trailing spaces were appearing in the import, and in some cases,
preventing correct matching when match is based off these fields.

Additionally I have added code to ensure culture and other fields are
trimmed. If these fields has leading or trailing spaces in the csv, they
where previously not being trimmed, and the term lookups would fail.

Custom column handlers had a similar issue and require trimming of
values to ensure leading and trailing spaces in the csv field do not
affect the import.