Add GA repository dimension, refs #12209
- Add config setting in 'app.yml' for repository dimension index:

Added explanatory comments and un-commented the GA settings for a
better understanding.

- Upgrade to Universal Analytics:

Custom dimensions are not available in 'ga.js', which is already
deprecated. See: https://support.google.com/analytics/answer/3450662

- Add GA repository dimension component:

Called from the index pages of IOs, repos and actors, checks the new
config setting, gets the resource's related repository and sets it's
source culture authorized form of name as the value of the dimension
before sending the page view hit.