Pass user to CSV import and sign in, refs #12083
Fixed issue where, when doing a CSV import from the web UI, via an asynchronous
job, information object audit logging wasn't noting the user that created the
job.

Added --user-id option to task for CSV import of information objects and logic
to sign the user in before import.

Added, to QubitCsvImport, the relaying of the current logged-in user to
the import task as a CLI option (if information objects are being imported).

 Conflicts:
	lib/task/import/csvImportTask.class.php