Fix /private redirect in viewAction, refs #12609
When AtoM is not installed in the default location, but is instead being served from a sub-directory of the root, it will have a URL like the following (rather the usual http://example.com/)

- http://example.com/path/to/atom/

In cases like this, when accessing an image/file in the "uploads/r/" folder however, the setResponseHeaders() method in viewAction.class.php will redirect to the absolute URL

- http://example.com/private/uploads/r/blah/blah/blah.jpg

instead of

- http://example.com/path/to/atom/private/uploads/r/blah/blah/blah.jpg

That is, portion of the URL path leading to the location of the installed AtoM instance ("/path/to/atom") is lost.  This will result in a 404 error.

This commit fixes the issue with setResponseHeaders() and ensures the URL path portion is retained.