Dockerfile change PHP version, refs #12821
Memcache will not install correctly from pecl as specified in our
Dockerfile under Alpine/PHP 7.3. This issue appears that it could be limited
to affecting Windows and Mac hosts and not Linux.

Update the Dockerfile to install Alpine with PHP 7.2.x. Add code to also
load libzip library which is necessary under PHP 7.3 but also works for
PHP 7.0 and up.