Digital object master discarding, refs #12840
Added functionality to the digital object load task, and model, to
support a new digital oject usage type: local files.

Local files exist outside of AtoM and, when associated with an
information object using the digital:object task, their representations
will be displayed on the information object's detail page but the
master file won't be accessible.