Nested set CLI task fix, refs #12754
The build-nested-set CLI task iterates over database rows in chunks of
4000. The current logic had an off-by-one error where if there were a
multiple of 4000 rows in the DB to iterate over, it would run one too
many times resulting in a fatal error.

I have changed the comparison operator so that it will not perform the
logic when the iterator is equal to the number of rows in the database.