Video player fixes, refs #12902
* Preload video metadata

This helps the MediaElement player to get the aspect ratio before the
user clicks on the play button scaling it better.

* Fix audio/video DO loading from the full-width treeview

This loads the mediaelement assets in the index of the information
object so that the player works when the DOs are loaded through AJAX
calls from interacting with the full-width treeview.

It also converts the mediaelementplayer setup into a Drupal Behavior
(like the previous flowplayer was) so that it can be reattached when
the DOM is updated.

* Fix video thumbnail generation

This fixes the generation of thumbnails for videos shorter than 30
seconds by calculating the half of the video using ffprobe.

It also unifies the thumbnail generation logic with the one used by
the Flash uploader.