Video player fixes after testing, refs #12902
* Load mediaelement JS assets last

When the user is not authenticated the mediaelement assets are loaded
after the jQuery library breaking the player functionality.

This sets the mediaelement JS assets loading in the 'last' position.

* Set fixed size for mediaelement player

This removes the responsive stretching mode from the mediaelement
player setup and sets a 320x240 size used previously by the
flowplayer.