Use config to generate jobs prefix, refs #13108
Add `workers_key` setting to `config/app.yml` and use it alongside the
`sf_root_dir` to generate the job prefix instead of the site title and
base URL.

The site title and base URL were used to generate the job prefix, which
leads to different prefixes when those values are changed in the
database or when using a different language. If the prefix changes
between the AtoM worker start and when a job is triggered, no worker
will be found to handle that job.

This prefix was introduced to avoid collisions between workers from
different instances connected to the same Gearman server. However, if
those instances are running on the same machine, the `sf_root_dir` will
be different on each of them. Therefore, using the title and URL only
prevents the collision when multiple instances are running on different
machines, on the same path and they're connected to the same Gearman
server.

This is not an usual scenario and requires to modify configuration files
manually, at least `config/gearman.yml` in some of those instances to
point to a different server. Therefore, setting a different value for
the `workers_key` on each instance seems like a good solution to avoid
the collisions.

In the majority of the deploys, even in those where multiple AtoM
instances are running on the same machine, this new setting can be left
empty.