Avoid double ancestors query on ACL, refs #13224
Calling count and then looping over the results of a criteria causes the
execution of a similar query twice, once for the count and another to
fetch the objects.

In this case, the count is not needed as the ancestors criteria is
followed by `andSelf()`, which will always return at least one result.