Don't use QubitTransactionFilter::getConnection()
The QubitTransactionFilter::getConnnection() is a Singleton wrapper around
Propel::getConnection() that doesn't include the optional arguments
supported by the latter method.

The name of QubitTransactionFilter::getConnection() can also give the mistaken
impression that the connection object returned by the method will have
different transaction behaviour than that returned by Propel::getConnection(),
which is not true (transactions are implemented in the symfony filter chain,
not in the connection object).