CSV import matching issue, refs #13439
This commit addresses an issue in Descriptions CSV import where the
secondary matching logic based on the description's title, identifier,
and repo was failing. The way the repository logic was being indexed
with a description was recently changed so that only specific fields
are being included instead of the entire QubitRepository object.

This commit allows an 'untouched' keyword field to be added for
translatable fields included in a '_partial_foreign_types' mapping.yml
section.

To fix the CSV import secondary matching, the information object's
repository partial_foreign_type section now specifies
'_rawFields: [authorizedFormOfName]' to allow keyword matching on repo
name. This feature can be used in any _partial_foreign_type.