AIP Download metadata only DIP, refs #13454
This commit fixes an issue where the 'Download file' link on the digital
object page was not working if the source of the digital object was a
metadata only DIP upload.

Now when a metadata only DIP upload is performed, the
relative_path_within_aip and aip_name will be saved as properties of the
object in AtoM's create digital object REST API endpoint. This
information is necessary when calling the AM Storage Service
'extract_file' API endpoint.