Fix missing treeview children, refs #13423
In a number of scenarios an Elasticsearch information object document
will not have a value for the "children" field, which was preventing
the document's children from being displayed in the full width
treeview.

To work around the problem an Elasticsearch query is used to count the
children of each node in the tree.  This workaround is significantly
slower for large trees, but all of the tree nodes are loaded correctly.