CSV Validator output changes, refs #13475
This commit includes formatting changes to the output report
- add two lines separating the summary at the top of the report from the
following test specific sections.
- format the test titles so the result is first and in all caps.
- Update column name check result so that the list of unknown column
names is returned with the non-verbose version of the output.
- Other csfixer code formatting changes.
- The language column validator is only available for information objects and
repositories.

Enhance the sample values test to highlight:
- duplicated columns
- empty columns

Column specific tests to check if a column is present or is duplicated
moved to the base validator. New php_unit tests added for the case where
a column is duplicated in the CSV file.