[C] Add Preservica metadata sync CLI task
This commit adds a new CLI task that syncs recently updated AtoM
descriptions to Preservica using the Preservica API for any
InformationObjects and DigitalObjects in AtoM with a corresponding
preservicaUUID property.

It also modifies the EAD template in AtoM slightly to use the XSD
schema expected by Preservica, and only sends single-level EAD.

When the --update-parents flag is used, the CLI task will attempt to
update parent descriptions as well as those for which AtoM has
Preservica UUIDs. This assumes that the hierarchies in AtoM and
Preservica are identical, as it works by walking up both from the
digital object simultaneously.