Treeview accordion enhancements
- Wrap fwtv in B5 accordion elements
- Refactored fullwidthtreeview.js to use js classes.

Since the treeview is always contained inside an accordion section now,
the fwtv setting text has been changed to indicate the setting causes
the fwtv to be closed by default on first load of a description page.

Old wording: 'Make full width treeview collapsible on description pages'
New: 'Make full width treeview collapsed on description pages'

Fixed css for fwtv ellipsis so that the left side is not pushed out
of the accordion.