Restart atom-worker job count, refs #xxxxx
Add new parameter to the `symfony jobs:worker` task to set the number of
jobs that will be run before the worker automatically exits. When set,
the worker will output log messages indicating what the job count is,
and what the max number of jobs is.

e.g. php symfony jobs:worker --max-job-count=10

If the worker task hits it's execution limit, it will return a code of
'111' which systemd can use to identify the reason the worker exited.