Treeview move index issue, refs #13567
This commit fixes an issue where full width treeview move no longer
functions due to the ES index (which populates the treeview) lft values
becoming out of sync with the database lft values when nodes are moved
via the UI. This action was not triggering re-index of affected nodes
resulting in treeview errors: 'Mismatch in current position' when the
database lft ordering did not match the ES index lft ordering. When this
occurs the only fix is to run a search:populate.

The range of treeview node objects that a move action encompasses need
to be re-indexed after a treeview move event. Since reindexing the
entire parent node could involve a large number of records, this fix
only reindexes those objects affected by the move, and performs a partial
update of lft values only instead of updating the entire ES document.

The performance of this fix is directly related to the number of nodes
affected by the move event - the 'further' the move, the more objects
will need reindexing, and the longer the move job will take. If a user
refreshes the treeview while the move job is re-indexing it is possible
for the UI ordering to appear out of sync temporarily with the DB.