Fix visibility bug for RAD notes, ref#13631
Change sfRadPlugin note visibility logic for
Rights and Conservation notes to compare term IDs
instead of strings to determine their visibility
status. This would ensure that it still functions
if English is not the source culture for those
strings.

Also add a __get method to QubitNote to handle
sourceCulture option. Currently these calls
go to BaseNote which ignores this option, and
this leads to hidden fields such as Rights or
Conservation showing up when language is changed.

See line 265-273 in
sfRadPlugin/templates/indexSuccess.php for a
scenario where the ['sourceCulture' => true]
is passed correctly but would being ignored.