Add error message for XLIFF file load (#1627)
Throwing an exception when Symfony runs into issues when loading XLIFF
files. This will help with catching broken XML files due to translation
errors. Without any error handling, this would lead Symfony to use an
empty array as the array of translations when it is unable to read XML
files, and would make it delete all existing translations for such a
file.