Fix issue nonce not assigned to GMaps style
Ensure that the CSP nonce is available in the page so that Google Maps
JS can use it. This eliminates CSP errors that occur when Google Maps are
included on a page, but the page does not have any other elements that
include the nonce. This change ensures the nonce appears at least once on
the page before the map component is displayed.